+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - `Edit Collection` TUI has been replaced by `Edit Recipe`, which opens the file to the selected recipe
   - This will make it much easier to make tweaks to a recipe
+- The `Body` and `Authentication` tabs of the `Recipe` pane are now disabled if the recipe doesn't have a body/authentication (respectively)
+- Disabled actions can no longer be selected in the action menu
 
 ## [4.1.0] - 2025-09-30